Move relay to its own module.
